added a way to buffer reading

+++ src/match.c
@@ -85,7 +85,7 @@ match* prev_match(match* match, list_t* matches)
 //changed inputs, return type
 void insert_at(match* newMatch, int index, list_t* matches)
 {
-    if (index > list_size(matches) - 1) 
+    if (index > (int) (list_size(matches) - 1)) 
     {
         list_append(matches, newMatch);
     }
